Move MailVerifier to application package
